requirements.txt
wheel
flask
gevent
gunicorn
grequests
pyyaml
cfscrape
bs4
prettytable
vulners
